requirements.txt
loguru
scipy
transformers
Pytorch > 1.10
